update to Blender 4.3

+++ source/blender/blenlib/BLI_hash.hh
@@ -66,7 +66,6 @@
 #include <utility>
 
 #include "BLI_string_ref.hh"
-#include "BLI_utildefines.h"
 
 namespace blender {
 
@@ -144,7 +143,8 @@ TRIVIAL_DEFAULT_INT_HASH(uint64_t);
 template<> struct DefaultHash<float> {
   uint64_t operator()(float value) const
   {
-    return *reinterpret_cast<uint32_t *>(&value);
+    /* Explicit `uint64_t` cast to suppress CPPCHECK warning.  */
+    return uint64_t(*reinterpret_cast<uint32_t *>(&value));
   }
 };
